requirements.txt
Django>=5.0.0
pymongo>=4.5.0
django-mongoengine>=0.4.0
python-dotenv>=1.0.0
pillow>=10.0.0
